[ADD] Example Of DLLs Folder VS
Added SFML Library as an example of DLL Files.
Additionally explained each part required within the comments.
Such as Including " Include Folder and Lib Folder "
And additionally Changing Working directory to start from  " DLLs "  instead of the root directory, this means that DLLs will be checked before any of the core files are and that states if to continue or not, avoiding complications of dependencies that do not exist.

Also see other parts of how it is copied in Build Events
Pre-Link Event Specifically.

Copy Statement.